BackendFastAPI/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary>=2.9.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-decouple==3.8
pydantic-settings==2.0.3

# Database & Validation
pydantic==2.5.0
email-validator==2.1.0

# Text Processing
python-slugify==8.0.1

# AWS Integration
boto3==1.34.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Image Processing
pillow==10.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
